+++ CHANGELOG.md
@@ -5,6 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.5] - 2023-09-18
+
+### Added
+
+- **Native**: Support for the `Integer` type for return codes.
+
+### Fixed
+
+- **Native**: Problems displaying recursive mode (argument number colliding with other calls).
+- **Native**: Fixed automatic detection of the number of arguments per function (experimental).
+
+### Changed
+
+- **Native**: Removed detection of `UUID` form in `hex` format.
+- **Crypto**: Removed detection of `UUID` form in `hex` format.
+
 ## [1.1.4] - 2023-08-15
 
 ### Added
@@ -68,6 +84,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 - Initial Release.
 
+[1.1.5]: https://github.com/hyugogirubato/Frida-CodeShare/releases/tag/v1.1.5
 [1.1.4]: https://github.com/hyugogirubato/Frida-CodeShare/releases/tag/v1.1.4
 [1.1.3]: https://github.com/hyugogirubato/Frida-CodeShare/releases/tag/v1.1.3
 [1.1.2]: https://github.com/hyugogirubato/Frida-CodeShare/releases/tag/v1.1.2